BACKUP 14.1.26/requirements.txt
# Arquivo: requirements.txt
appium-python-client
selenium
pytest
pytest-dotenv
allure-pytest